Separated the button class into a new file. Updated the project inclusion accordingly.
